+++ fbw-a32nx/src/systems/instruments/src/EFB/Ground/Pages/Payload/A380_842/Seating/SeatMapWidget.tsx
@@ -1,13 +1,13 @@
-// Copyright (c) 2021-2023 FlyByWire Simulations
-//
-// SPDX-License-Identifier: GPL-3.0
-
 import React, { useEffect, useRef, useState } from 'react';
-import { BitFlags, usePersistentProperty } from '@flybywiresim/fbw-sdk';
+import { BitFlags } from '@shared/bitFlags';
 import * as ReactDOMServer from 'react-dom/server';
-import { CanvasConst, SeatConstants, SeatInfo, PaxStationInfo, TYPE, RowInfo } from './Constants';
-import { Seat } from '../../../../Assets/Seat';
-import { SeatOutlineBg } from '../../../../Assets/SeatOutlineBg';
+import { usePersistentProperty } from '@instruments/common/persistence';
+import { AirplaneFill } from 'react-bootstrap-icons';
+import { SelectGroup, SelectItem } from '../../../../../UtilComponents/Form/Select';
+import { CanvasConst, SeatConstants, SeatInfo, PaxStationInfo, TYPE, RowInfo } from '../../Seating/Constants';
+import { Seat } from '../../../../../Assets/Seat';
+import { A380SeatOutlineBg } from '../../../../../Assets/A380SeatOutlineBg';
+// import { t } from '../../../../../translation';
 
 interface SeatMapProps {
     seatMap: PaxStationInfo[],
@@ -108,6 +108,7 @@ export const SeatMapWidget: React.FC<SeatMapProps> = ({ seatMap, desiredFlags, a
                 let seatId = 0;
                 for (let row = 0; row < seatMap[station].rows.length; row++) {
                     xOff = addXOffsetRow(xOff, station, row);
+                    console.log(`drawRow(xOff: ${xOff}, station: ${station}, row: ${row}, seatID: ${seatId})`);
                     drawRow(xOff, station, row, seatMap[station].rows[row], seatId);
                     seatId += seatMap[station].rows[row].seats.length;
                 }
@@ -203,8 +204,22 @@ export const SeatMapWidget: React.FC<SeatMapProps> = ({ seatMap, desiredFlags, a
 
     return (
         <div className="flex relative flex-col">
-            <SeatOutlineBg stroke={getTheme(theme)[0]} highlight="#69BD45" />
+            <A380SeatOutlineBg stroke={getTheme(theme)[0]} highlight="#69BD45" />
             <canvas className="absolute cursor-pointer" ref={canvasRef} style={{ transform: `translateX(${canvasX}px) translateY(${canvasY}px)` }} />
+
+            <div className="flex absolute top-full flex-row px-4 w-fit">
+                <div><AirplaneFill size={25} className="my-1 mx-3" /></div>
+                <SelectGroup>
+                    <SelectItem
+                        selected
+                    >
+                        Main
+                    </SelectItem>
+                    <SelectItem>
+                        Upper
+                    </SelectItem>
+                </SelectGroup>
+            </div>
         </div>
     );
 };
