+++ fbw-common/src/systems/instruments/src/react/simVars.tsx
@@ -76,6 +76,92 @@ export const useSimVar = (
 };
 
 /**
+ * The useSimVarList hook provides an easy way to read and write a list of SimVars from React.
+ * This is primarily designed for use with dynamic lists.
+ *
+ * It's signature is similar to useState and it regularly refreshes the SimVar
+ * to ensure your React component stays in sync with the SimVar being modified
+ * from outside your component (like from other components, XML or SimConnect).
+ *
+ * You may optionally specify the refresh interval. If the same SimVar
+ * is used in multiple places, this hook will automatically deduplicate those
+ * for maximum performance, rather than fetching the SimVar multiple times.
+ * Setting the SimVar will instantly cause it to be updated in all other places
+ * within the same React tree.
+ *
+ * @param names A list of the name of the SimVars.
+ * @param units A list of the units of the SimVars.
+ * @param refreshInterval The time in milliseconds that needs to elapse before
+ * the next render will cause a SimVar refresh from the simulator.
+ *
+ * @example
+ * // the return value is the value itself and a setter, similar to useState
+ * const [v1, setV1] = useSimVar('L:AIRLINER_V1_SPEED', 'Knots');
+ *
+ * @example
+ * // only refresh the SimVar every 500ms
+ * const [lightsTest] = useSimVar('L:A32NX_OVHD_INTLT_ANN', 'Bool', 500);
+ *
+ * @returns {[*, (function(*): void)]}
+ *
+ * @see {@link useSplitSimVar} if your SimVar is set through a K event
+ * @see {@link useInteractionSimVar} if you emit an H event whenever your SimVar changes
+ * @see {@link useGlobalVar} if you have a Global Var instead
+ */
+export const useSimVarList = (
+    names: string[],
+    units: UnitName[],
+    refreshInterval = 0,
+): [SimVarValue, (newValueOrSetter: SimVarValue | SimVarSetter
+) => void] => {
+    const lastUpdate = useRef(Date.now() - refreshInterval - 1);
+
+    const [stateList, setStateList] = useState(() => {
+        const simVars: number[] = [];
+        names.forEach((name, index) => {
+            simVars.push(SimVar.GetSimVarValue(name, units[index]));
+        });
+        return simVars;
+    });
+
+    const updateCallback = useCallback(() => {
+        const delta = Date.now() - lastUpdate.current;
+
+        if (delta >= refreshInterval) {
+            lastUpdate.current = Date.now();
+            names.forEach((name, index) => {
+                stateList[index] = SimVar.GetSimVarValue(name, units[index]);
+            });
+            setStateList(stateList);
+        }
+    }, [
+        names, ...names,
+        units, ...units,
+        refreshInterval,
+    ]);
+
+    useUpdate(updateCallback);
+
+    const setter = useCallback((valuesOrSetter: any | SimVarSetter) => {
+        const executedValues = typeof valuesOrSetter === 'function' ? valuesOrSetter(stateList) : valuesOrSetter;
+
+        executedValues.forEach((executedValue, index) => {
+            SimVar.SetSimVarValue(names[index], units[index], executedValue);
+            stateList[index] = executedValue;
+        });
+        setStateList(stateList);
+
+        return stateList;
+    }, [
+        names, ...names,
+        units, ...units,
+        ...stateList, stateList,
+    ]);
+
+    return [stateList, setter];
+};
+
+/**
  * The useGlobalVar hook provides an easy way to read and write GlobalVars from
  * React. The signature is similar to useSimVar, except for the return being a
  * single value as it is non-writeable.
